Fix incompatibility with PWG Stuffs
The modus theme displays a "..." link on mobile devices, which display a
list of actions by clicking on it.
The JS code assumes that there is only one .categoryActions element in
the body. If the user is logged in as administrator, and 'stuffs'  have
been added to the page, the selector is wrong: it will return more than 1
element.

This patch takes care of selecting the .categoryActions element next to the
link.